refactor: extract admin order move endpoint into dedicated router
